Populating reviews and review-based info complete
Utilized (idk who made this) the backend endpoint to fetch reviews for a club by a specific id in the club page through useEffect, passing in the clubId given by the URL.

The result is stored in a reviews object, which I iterated over to produce the average overall rating, application status, time commitment, and most common major with O(N) time complexity. These computed values are passed into the corresponding ClubInfoCard to be displayed as general information for the club.

Refactored review card component to take in the following fields and populate accordingly: description, major, application, time, position, rating, and date. These fields are mapped for each review in the reviews object. Updated the username field of each review to instead say the date, styled to be more subtle with minimalistic formatting.